refactor: extract BackupDatabaseAdapter to break the backup_target import cycle (#300)
