+++ lib/features/backup/data/services/backup_target.dart
@@ -1,7 +1,7 @@
 import 'package:path/path.dart' as p;
 
+import 'package:submersion/features/backup/data/services/backup_database_adapter.dart';
 import 'package:submersion/features/backup/data/services/backup_saf_port.dart';
-import 'package:submersion/features/backup/data/services/backup_service.dart';
 
 /// True iff [ref] is a SAF document/tree URI rather than a filesystem path.
 /// Only Android's picker produces these, so it doubles as the platform branch:
